+++ example-app/lib/express/saml.js
@@ -7,6 +7,8 @@ const passport = require('passport')
 const SamlStrategy = require('passport-saml').Strategy
 // const fs = require('fs')
 
+const { SAML_ISSUER, SAML_ENTRYPOINT, SAML_CALLBACK_URL } = global.CONFIG
+
 module.exports = function(app) {
   app.use(passport.initialize())
 
@@ -16,9 +18,9 @@ module.exports = function(app) {
   passport.use('saml', new SamlStrategy(
     {
       // cert: fs.readFileSync('/path/to/adfs.acme_tools.com.crt', 'utf-8'), // PEM in single string to ensure ADFS Server is not a fake one
-      issuer: 'saml-poc',
-      callbackUrl: 'http://127.0.0.1:3000/api/saml/login/callback', // 'https://acme_tools.com/adfs/postResponse',
-      entryPoint: 'http://127.0.0.1:8081/simplesaml/saml2/idp/SSOService.php', // 'https://adfs.acme_tools.com/adfs/ls/',
+      issuer: SAML_ISSUER,
+      callbackUrl: SAML_CALLBACK_URL, // 'https://acme_tools.com/adfs/postResponse',
+      entryPoint: SAML_ENTRYPOINT, // 'https://adfs.acme_tools.com/adfs/ls/',
       // privateCert: fs.readFileSync('/path/to/acme_tools_com.key', 'utf-8'), // not needed yet
       // authnContext: 'http://schemas.microsoft.com/ws/2008/06/identity/authenticationmethod/password',
       // authnContext: 'http://schemas.microsoft.com/ws/2008/06/identity/authenticationmethod/windows',