+++ example-vite/src/views/SignIn.vue
@@ -8,6 +8,7 @@
         <mwc-textfield label="Password" outlined type="password" v-model="password"></mwc-textfield>
         <div class="buttons-box-flex">
           <mwc-button raised label="Login" type="button" @click="login"></mwc-button>
+          <mwc-button raised label="SAML" type="button" @click="samlLogin"></mwc-button>
           <!-- <o-button @click="login">Login Use Oruga UI</o-button> -->
         </div>
         <p><router-link to="/signup">Sign Up</router-link></p>
@@ -27,13 +28,16 @@
 <script>
 import { ref, onMounted, onBeforeUnmount, onUnmounted } from 'vue'
 import { useStore } from 'vuex'
+import { useRoute } from 'vue-router'
+
 
 import { useXhr } from '/src/plugins/xhr.js'
 import { useI18n } from '/src/plugins/i18n.js'
 import { useWs } from '/src/plugins/ws.js'
 
-import apollo from '/src/lib/esm/apollo' // may not need to use provide/inject if no reactivity ?
-import { DO_HELLO } from '/src/queries'
+import { samlLogin } from '/src/lib/esm/saml.js'
+import apollo from '/src/lib/esm/apollo.js' // may not need to use provide/inject if no reactivity ?
+import { DO_HELLO } from '/src/queries.js'
 
 const apolloClient = apollo.get()
 if (apolloClient) {
@@ -51,6 +55,8 @@ if (apolloClient) {
 export default {
   setup(props, context) {
     const store = useStore()
+    const route = useRoute()
+
     const http = useXhr()
     const i18n = useI18n()
     const ws = useWs()
@@ -75,7 +81,7 @@ export default {
 
     onUnmounted(() => console.log('signIn unmounted'))
     onMounted(async () => {
-      console.log('signIn mounted!')
+      console.log('signIn mounted!', route.hash) // deal with hashes here if necessary
 
       setToLogin()
       errorMessage.value = ''
@@ -167,6 +173,7 @@ export default {
     }
 
     return {
+      samlLogin,
       email, // data
       password,
       errorMessage,