+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Version 0.4.4
 - [chore] update packages & cleanup & work on improving documentation
 - [nobundler-example] created initial bwc-autocomplete (based on bulma)
+- [example-all] implement saml callback and configs
 - [inprogress]
   - [chore] use monorepo when npm 7 is released ?
   - [example-app] improve vanilla app